requirements.txt
wxpython
pillow
google-cloud-translate
google-cloud-vision
gtts